Removed v1.1 until after a PROD deployment
