+++ src/backend/oracle-data-api/src/main/java/ca/bc/gov/open/jag/tco/oracledataapi/controller/v1_0/DisputeController.java
@@ -1,7 +1,11 @@
 package ca.bc.gov.open.jag.tco.oracledataapi.controller.v1_0;
 
+import java.util.Date;
 import java.util.List;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -11,13 +15,25 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import ca.bc.gov.open.jag.tco.oracledataapi.error.DeprecatedException;
 import ca.bc.gov.open.jag.tco.oracledataapi.model.Dispute;
+import ca.bc.gov.open.jag.tco.oracledataapi.model.DisputeStatus;
+import ca.bc.gov.open.jag.tco.oracledataapi.service.DisputeService;
+import ca.bc.gov.open.jag.tco.oracledataapi.service.LookupService;
 import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
 
 @RestController(value = "DisputeControllerV1_0")
 @RequestMapping("/api/v1.0")
 public class DisputeController {
+
+	@Autowired
+	private DisputeService disputeService;
+
+	@Autowired
+	private LookupService lookupService;
+
+	private Logger log = LoggerFactory.getLogger(DisputeController.class);
 
 	/**
 	 * GET endpoint that retrieves all the dispute detail from the database
@@ -25,21 +41,20 @@ public class DisputeController {
 	 * @return list of all dispute tickets
 	 */
 	@GetMapping("/disputes")
-	@Deprecated
 	public List<Dispute> getAllDisputes() {
-		throw new DeprecatedException();
+		log.debug("Retrieve all disputes endpoint is called." + new Date());
+		return disputeService.getAllDisputes();
 	}
 
 	/**
 	 * GET endpoint that retrieves the detail of a specific dispute
 	 *
-	 * @param disputeId
+	 * @param id
 	 * @return {@link Dispute}
 	 */
-	@GetMapping("/dispute/{disputeId}")
-	@Deprecated
-	public Dispute getDispute(@PathVariable("disputeId") int disputeId) {
-		throw new DeprecatedException();
+	@GetMapping("/dispute/{id}")
+	public Dispute getDispute(@PathVariable Integer id) {
+		return disputeService.getDisputeById(id);
 	}
 
 	/**
@@ -47,10 +62,9 @@ public class DisputeController {
 	 *
 	 * @param id of the {@link Dispute} to be deleted
 	 */
-	@DeleteMapping("/dispute/{disputeId}")
-	@Deprecated
-	public void deleteDispute(@PathVariable("disputeId") int disputeId) {
-		throw new DeprecatedException();
+	@DeleteMapping("/dispute/{id}")
+	public void deleteDispute(@PathVariable Integer id) {
+		disputeService.delete(id);
 	}
 
 	/**
@@ -60,31 +74,78 @@ public class DisputeController {
 	 * @return id of the saved {@link Dispute}
 	 */
 	@PostMapping("/dispute")
-	@Deprecated
-	public int saveDispute(@RequestBody Dispute dispute) {
-		throw new DeprecatedException();
+	public Integer saveDispute(@RequestBody Dispute dispute) {
+		disputeService.save(dispute);
+		return dispute.getId();
 	}
 
 	/**
-	 * PUT endpoint that updates the dispute detail
+	 * PUT endpoint that updates the dispute detail, setting the status to REJECTED.
 	 *
 	 * @param dispute to be updated
+	 * @param id of the saved {@link Dispute} to update
 	 * @return {@link Dispute}
 	 */
-	@PutMapping("/dispute")
-	@Deprecated
-	public Dispute update(@RequestBody Dispute dispute) {
-		throw new DeprecatedException();
+	@Operation(summary = "Updates the status of a particular Dispute record to REJECTED.")
+	@ApiResponses({
+		@ApiResponse(responseCode = "200", description = "Ok"),
+		@ApiResponse(responseCode = "400", description = "Bad Request."),
+		@ApiResponse(responseCode = "404", description = "Dispute record not found. Update failed.")
+	 //,@ApiResponse(responseCode = "405", description = "Only NEW statuses can be changed to REJECTED. Update failed.") // Not currently a requirement.
+	})
+	@PutMapping("/dispute/{id}/reject")
+	public void rejectDispute(@PathVariable Integer id, @RequestBody String rejectedReason) {
+		disputeService.setStatus(id, DisputeStatus.REJECTED, rejectedReason);
+	}
+
+	/**
+	 * PUT endpoint that updates the dispute detail, setting the status to CANCELLED.
+	 *
+	 * @param dispute to be updated
+	 * @param id of the saved {@link Dispute} to update
+	 * @return {@link Dispute}
+	 */
+	@Operation(summary = "Updates the status of a particular Dispute record to CANCELLED.")
+	@ApiResponses({
+		@ApiResponse(responseCode = "200", description = "Ok"),
+		@ApiResponse(responseCode = "400", description = "Bad Request."),
+		@ApiResponse(responseCode = "404", description = "Dispute record not found. Update failed.")
+     //,@ApiResponse(responseCode = "405", description = "Only NEW statuses can be changed to CANCELLED. Update failed.") // Not currently a requirement.
+	})
+	@PutMapping("/dispute/{id}/cancel")
+	public void cancelDispute(@PathVariable Integer id) {
+		disputeService.setStatus(id, DisputeStatus.CANCELLED);
+	}
+
+	/**
+	 * PUT endpoint that updates the dispute detail, setting the status to PROCESSING.
+	 *
+	 * @param dispute to be updated
+	 * @param id of the saved {@link Dispute} to update
+	 * @return {@link Dispute}
+	 */
+	@Operation(summary = "Updates the status of a particular Dispute record to PROCESSING.")
+	@ApiResponses({
+		@ApiResponse(responseCode = "200", description = "Ok"),
+		@ApiResponse(responseCode = "400", description = "Bad Request."),
+		@ApiResponse(responseCode = "404", description = "Dispute record not found. Update failed.")
+	 //,@ApiResponse(responseCode = "405", description = "Only NEW statuses can be changed to PROCESSING. Update failed.") // Not currently a requirement.
+	})
+	@PutMapping("/dispute/{id}/submit")
+	public void submitDispute(@PathVariable Integer id) {
+		disputeService.setStatus(id, DisputeStatus.PROCESSING);
 	}
 
 	/**
 	 * GET endpoint that refreshes all codetables cached in redis
 	 */
 	@GetMapping("/codetable/refresh")
-	@Operation(summary = "An endpoint hook to trigger a redis rebuild of cached codetable data.", description = "The codetables in redis are cached copies of data pulled from Oracle to ensure TCO remains stable. This data is periodically refreshed, but can be forced by hitting this endpoint.")
-	@Deprecated
+	@Operation(
+			summary = "An endpoint hook to trigger a redis rebuild of cached codetable data.",
+			description = "The codetables in redis are cached copies of data pulled from Oracle to ensure TCO remains stable. This data is periodically refreshed, but can be forced by hitting this endpoint."
+			)
 	public void codeTableRefresh() {
-		throw new DeprecatedException();
+		lookupService.refresh();
 	}
 
 }
